+++ src/dspf-edit.commands/register-commands.ts
@@ -22,6 +22,7 @@ import { addCommandsRecord } from './dspf-edit.add-commands-record';
 import { addColor } from './dspf-edit.add-color';
 import { addAttribute } from './dspf-edit.add-attribute';
 import { addKeyCommand } from './dspf-edit.add-keys';
+import { addSflDropSflFold } from './dspf-edit.sfldrop-sflfold';
 import { addValidityCheck } from './dspf-edit.add-validity-check';
 import { editingKeywords } from './dspf-edit.add-editing-keywords';
 import { addErrorMessage } from './dspf-edit.add-error-messages';
@@ -59,6 +60,7 @@ export const commands = [
     { name: 'addColor', handler: addColor, needsTreeProvider: false },
     { name: 'addAttribute', handler: addAttribute, needsTreeProvider: false },
     { name: 'addKey', handler: addKeyCommand, needsTreeProvider: false },
+    { name: 'addSflDropSflFold', handler: addSflDropSflFold, needsTreeProvider: false },
     { name: 'addValidityCheck', handler: addValidityCheck, needsTreeProvider: false },
     { name: 'addEditingKeywords', handler: editingKeywords, needsTreeProvider: false },
     { name: 'addErrorMessage', handler: addErrorMessage, needsTreeProvider: false },